RUE-532: constant u64 indices above i64::MAX get the compile-time bounds check (spec 7.1:9)
try_get_const_index_checked narrowed the evaluator's i128 through
ConstValue::as_integer() -> Option<i64>, so a u64 constant above i64::MAX
returned None — an exactly-known out-of-bounds index was indistinguishable
from a runtime index and deferred to the runtime trap (verified: exit 101
instead of E0902). The query now returns the full i128, the four bounds
checks compare in i128, and IndexOutOfBounds carries i128 so the diagnostic
reports the exact index (18446744073709551615, not a clamp).

Tests: 7.1:9 spec cases — huge-constant rejection pinning the exact index in
the message, and an in-bounds u64-constant control.

+++ crates/rue-air/src/sema/analysis/instructions.rs
@@ -684,7 +684,7 @@ impl<'a> Sema<'a> {
             // index's resolved operand types so an overflowing index expression
             // is a compile-time error, not a folded runtime panic (RUE-234).
             if let Some(const_index) = self.try_get_const_index_checked(index, ctx)? {
-                if const_index < 0 || const_index as u64 >= array_len {
+                if const_index < 0 || const_index >= array_len as i128 {
                     return Err(CompileError::new(
                         ErrorKind::IndexOutOfBounds {
                             index: const_index,
